chore: add init parameters for client
